Support for building with Remix 2.0.0 (#151)
No big changes required, it seems. The main one affecting this project
is that route.handle is now of type "unknown", so I added some
additional casts in order to avoid type checking errors.

In addition to this, Remix requires Typescript 5.1.0 or later. This, in
turn, requires upgrading @typescript-eslint to at least 5.61.0 (I went
with 5.62.0 since that's the last version before 6.0.0).

Of course, just running `npm install` means we get a bunch of other
packages upgraded. I just went with it since none of the packages seem
like they would cause breakage on upgrade.